d/p/CVE-2024-7264-1.patch: backport patch
In this version of cURL, GTime2str doesn't return CURLcode, so change
the error return to NULL. It also don't have a dynbuf as argument, so
remove it from Curl_x509_GTime2str helper function. This function is
only used for unit tests and although we don't enable it, leaving it
with the wrong return type and arguments could cause FTBFS in case
anyone wants to rebuild this package with --enable-debug.